requirements.txt
google-genai>=0.3.0
google-adk>=0.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
gradio>=4.0.0
pillow>=10.0.0
gtts>=2.5.0